Refactor code - 리뷰 반영 - EventTranslator 를 app 으로 이동 - Account-application 에 외부 모듈 의존성 제거
