+++ common/src/main/java/kr/flab/movieon/common/ApiResponse.java
@@ -16,10 +16,6 @@ public final class ApiResponse<T> {
         this.message = message;
     }
 
-    public static <T> ApiResponse<T> success() {
-        return new ApiResponse<>(true, null, null);
-    }
-
     public static <T> ApiResponse<T> success(T data) {
         return new ApiResponse<>(true, data, null);
     }
